Refactor code completely
