+++ data_cascade/loader.py
@@ -1,20 +1,21 @@
-
 """Top-level API to load a data cascade from a root directory."""
 
 from __future__ import annotations
+
 from pathlib import Path
-from typing import Any, Dict, Tuple
+from typing import Any, Dict
+
 from .config import SUPPORTED_EXTS_DEFAULT, ensure_dir
-from .logging_utils import get_logger
-from .traverse import load_directory_node
-from .mapping import CascadeMap
-
 # import handlers to register
-from . import handlers_yaml  # noqa: F401
-from . import handlers_json  # noqa: F401
-from . import handlers_toml  # noqa: F401
+from .handlers import json  # noqa: F401
+from .handlers import toml  # noqa: F401
+from .handlers import yaml  # noqa: F401
+from .logging_utils import get_logger
+from .mapping import CascadeMap
+from .traverse import load_directory_node
 
-logger = get_logger("loader")
+log = get_logger(__name__)
+
 
 def load_data_cascade(
     root: Path | str,
@@ -23,7 +24,7 @@ def load_data_cascade(
 ) -> tuple[Dict[str, Any], CascadeMap]:
     root_path = Path(root)
     ensure_dir(root_path)
-    logger.info("Loading data cascade from %s", root_path)
+    log.info("Loading data cascade from %s", root_path)
     data, cmap = load_directory_node(root_path, allowed_exts=allowed_exts)
-    logger.info("Finished loading cascade from %s", root_path)
+    log.info("Finished loading cascade from %s", root_path)
     return data, cmap
