+++ data_cascade/merge/__init__.py
@@ -1,0 +1,2 @@
+from .merge import merge_values
+from .strategy import DictMode, ListMode, MergeStrategy
