+++ data_cascade/pathops/__init__.py
@@ -1,4 +1,4 @@
+from .access import delete_at, get_at, is_int_segment, set_at
+from .parse import join_path, parse_path
 
-from .parse import parse_path, join_path
-from .access import get_at, set_at, delete_at, is_int_segment
 __all__ = ["parse_path", "join_path", "get_at", "set_at", "delete_at", "is_int_segment"]
