+++ tests/test_roundtrip.py
@@ -1,9 +1,12 @@
-
 from __future__ import annotations
-from pathlib import Path
+
 import json
+from pathlib import Path
+
 import pytest
+
 from data_cascade import load_data_cascade, save_data_cascade
+
 
 @pytest.fixture()
 def sample_tree(tmp_path: Path) -> Path:
@@ -12,9 +15,12 @@ def sample_tree(tmp_path: Path) -> Path:
     (root / "__main__.yaml").write_text("name: Alpha\nversion: 1\n", encoding="utf-8")
     (root / "team.yaml").write_text("members:\n  - Alice\n  - Bob\n", encoding="utf-8")
     (root / "team").mkdir()
-    (root / "team" / "roles.yaml").write_text("roles:\n  Alice: Lead\n  Bob: Dev\n", encoding="utf-8")
-    (root / "numbers.json").write_text(json.dumps([1,2,3]), encoding="utf-8")
+    (root / "team" / "roles.yaml").write_text(
+        "roles:\n  Alice: Lead\n  Bob: Dev\n", encoding="utf-8"
+    )
+    (root / "numbers.json").write_text(json.dumps([1, 2, 3]), encoding="utf-8")
     return root
+
 
 def test_roundtrip_modify_add_delete(sample_tree: Path):
     data, cmap = load_data_cascade(sample_tree)
@@ -25,6 +31,7 @@ def test_roundtrip_modify_add_delete(sample_tree: Path):
     save_data_cascade(sample_tree, data, cmap)
     data2, _ = load_data_cascade(sample_tree)
     assert data2 == data
+
 
 def test_new_root_key_goes_to_main(sample_tree: Path):
     data, cmap = load_data_cascade(sample_tree)
